refactor layer version picker into a separate component
